+++ CHANGES.rst
@@ -4,3 +4,5 @@ Version 1.0.0
 Unreleased
 
 - Initial Release
+
+-  Updated null values handler to remove exclusive choosing of a null values handling method. https://github.com/preprocessy/preprocessy/issues/100